+++ example/entities/random_mover.gd
@@ -1,3 +1,5 @@
 @tool
 class_name RandomMover
 extends Entity
+
+@onready var color = Color(randf(), randf(), randf())