+++ example/main.gd
@@ -11,5 +11,6 @@ func _ready() -> void:
 func _process(delta: float) -> void:
 	world.process(delta, 'gameplay')
 
+
 func _physics_process(delta: float) -> void:
 	world.process(delta, 'physics')